% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{WFRD}
\alias{WFRD}
\title{Perform Robust Inference on Fuzzy RD Coefficients when identification is weak as detailed in Feir, Lemieux, Marmer (2016)}
\usage{
WFRD(
  Y,
  X,
  Z,
  controls = NULL,
  threshold,
  bandwidth = NULL,
  kernel = NULL,
  alpha = 0.05,
  beta0 = 0,
  stand_inf = TRUE,
  bw_select = "rdrobust"
)
}
\arguments{
\item{Y}{dependent varaible}

\item{X}{Treatment Variable}

\item{Z}{Running variable that affects treatment probability}

\item{controls}{Variables to control for}

\item{threshold}{the threshold for the discontinuity}

\item{bandwidth}{the bandwidth for the kernel}

\item{kernel}{the kernel function used}

\item{alpha}{the significance level}

\item{beta0}{the null hypothesis}

\item{stand_inf}{whether standard inference is included or not}

\item{bw_select}{"rdrobust" (default) for using rdbwselect from rdrobust package. "silverman" for silverman rule of thumb.}
}
\value{
A list object with important variables
}
\description{
This function estimates a Fuzzy RD model and performs robust inference in the case where the magnitude of the effect is small (weak identification)
requires a formula, a dataframe, a threshold, a bandwidth (default is Silverman's Rule of Thumb), a kernel (defualt is Uniform),
a significance level (default is .05), a null hypothesis value (default is 0), and an option to include standard inference approaches (default is FALSE)
}
