% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{WFRD_multiple_bw}
\alias{WFRD_multiple_bw}
\title{Perform Robust Inference on Fuzzy RD Coefficients when identification is weak as detailed in Feir, Lemieux, Marmer (2016) for many bandwidths.}
\usage{
WFRD_multiple_bw(
  Y,
  X,
  Z,
  controls = NULL,
  threshold,
  bandwidths,
  kernel = "uniform",
  alpha = 0.05,
  beta0 = 0,
  stand_inf = TRUE,
  plot_bool = TRUE,
  plot_stand = TRUE
)
}
\arguments{
\item{Y}{dependent varaible}

\item{X}{Treatment Variable}

\item{Z}{Running variable that affects treatment probability}

\item{controls}{Variables to control for}

\item{threshold}{the threshold for the discontinuity}

\item{bandwidths}{a list of bandwidths used for the kernel}

\item{kernel}{the kernel function used}

\item{alpha}{the significance level}

\item{beta0}{the null hypothesis}

\item{stand_inf}{whether standard inference is included or not}

\item{plot_bool}{whether a plot should be provided}

\item{plot_stand}{whether the standard confidence intervals should be plotted as well}
}
\value{
A list object with important variables and a plot
}
\description{
This function estimates a Fuzzy RD model and performs robust inference for many bandwidths in the case where the magnitude of the effect is small (weak identification)
requires a formula, a dataframe, a threshold, a list of bandwidths, a kernel (defualt is Uniform),
a significance level (default is .05), a null hypothesis value (default is 0), an option to include standard inference approaches (default is FALSE), and
a boolean that indicates whether a plot with bandwidths on the x axis and confidence sets on the y axis should be displayed (default is TRUE).
}
